Refactor: reorganize dir structure
