+++ client/src/comps/form-comps/FormButton.ts
@@ -1,4 +1,4 @@
-import Component from "./Component"
+import Component from "../Component"
 
 export default class FormButton extends Component {
 	constructor(
