+++ client/src/comps/form-comps/FormCheckbox.ts
@@ -1,4 +1,4 @@
-import Component from "./Component"
+import Component from "../Component"
 
 export default class FormCheckbox extends Component {
 	private readonly input: HTMLInputElement
