+++ client/src/comps/form-comps/FormMultiSelect.ts
@@ -1,4 +1,4 @@
-import Component from "./Component"
+import Component from "../Component"
 
 export type FormMultiSelectOption = {
 	id: string
