+++ client/src/comps/form-comps/FormTextInput.ts
@@ -1,5 +1,5 @@
-import Component from "./Component"
-import EmojiButton from "./EmojiButton"
+import Component from "../Component"
+import EmojiButton from "../EmojiButton"
 
 export default class FormTextInput extends Component {
 	private readonly input: HTMLInputElement
