+++ client/src/main.ts
@@ -3,13 +3,13 @@ import "../md.css"
 import "highlight.js/styles/github-dark.min.css"
 import App from "./comps/App"
 import ThreadletAPI from "@j0code/threadlet-api/v0"
-import PostView from "./comps/PostView"
+import PostView from "./comps/views/PostView"
 import { User } from "matrix-js-sdk"
 import Login from "./comps/Login"
-import RoomView from "./comps/RoomView"
-import RoomCreationForm from "./comps/RoomCreationForm"
-import RoomInvitationForm from "./comps/RoomInvitationForm"
-import RoomInviteForm from "./comps/RoomInviteForm"
+import RoomView from "./comps/views/RoomView"
+import RoomCreationForm from "./comps/forms/RoomCreationForm"
+import RoomInvitationForm from "./comps/forms/RoomInvitationForm"
+import RoomInviteForm from "./comps/forms/RoomInviteForm"
 
 export let api: ThreadletAPI
 export let app: App
